\name{BHS}
\alias{BHS}
\docType{data}
\title{
Busselton Health Study
}
\description{
This dataset is a subset of the Busselton Health study. The Busselton Health study was a repeated cross-sectional survey that was conducted to the community of Busselton in Western Australian.
}
\usage{data(BHS)}
\format{
  A data frame with 100 observations on the following 18 variables.
  \describe{
    \item{\code{PAIR}}{spouse pair id number}
    \item{\code{AGE}}{age at survey}
    \item{\code{SEX}}{sex}
    \item{\code{SBP}}{systolic blood pressure}
    \item{\code{DBP}}{diastolic blood pressure}
    \item{\code{BMI}}{body mass index}
    \item{\code{CHOL}}{cholesterol level}
    \item{\code{DIABETES}}{history of diabetes}
    \item{\code{RXHYPER}}{on blood pressure treatment}
    \item{\code{CHID}}{history of coronary heart disease}
    \item{\code{SMOKE}}{smoking status}
    \item{\code{DRINKING}}{alcohol consumption level}
    \item{\code{SURVTIME}}{survival time from survey data to date last known alive}
    \item{\code{DTHCENS}}{censoring indicator}
    \item{\code{CHDCENS}}{indicator of the death from coronary heart disease }
    \item{\code{CVDCENS}}{indicator of the death from cardiovascular disease}
    \item{\code{SMOKE1}}{indicator of ex-smoker}
    \item{\code{SMOKE2}}{indicator of current smoker}  
    }
}

\details{
This dataset is a subset of the Busselton Health study. The Busselton Health study was a repeated cross-sectional survey that was conducted to the community of Busselton in Western Australian.
}
\source{
He, W., Yi, G. Y. and Xiong, J. (2007). Accelerated Failure Time Models with Covariates Subject to Measurement Error. \emph{Statistics in Medicine}, 26, 4817-4832.

Knuiman, M. W., Cullent, K. J., Bulsara, M. K., Welborn, T. A. and Hobbs, M. S. T. (1994). Mortality trends, 1965 to 1989, in Busselton, The Site of Repeated Health Surveys and Interventions. \emph{Australian Journal of Public Health}, 18, 129-135.
}

\seealso{\code{\link[simexaft]{simexaft}}}

\keyword{datasets}


