\name{print.simexaft}
\alias{print.simexaft}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
 Print Method for the SIMEXAFT Class
}
\description{
Printing the most important values in a clear way. 
}
\usage{
\method{print}{simexaft}(x, digits = max(3, getOption("digits") - 3), \dots)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ object of class SIEMXAFT. }
  \item{digits}{ number of digits to be printed. }
  \item{\dots}{ arguments passed to other functions. }
}
\references{
Genz, A., Bretz, F., Miwa, T., Mi, X., Leisch, F., Scheipl, F. and Hothorn, T. (2011). mvtnorm: Multivariate Normal and t Distributions. R package version 0.9-9991, URL http://CRAN.
R-project.org/package=mvtnorm.

He, W., Yi, G. Y. and Xiong, J. (2007). Accelerated Failure Time Models with Covariates Subject to Measurement Error. \emph{Statistics in Medicine}, 26, 4817-4832.

Therneau, T. and Lumley, T. (2011). survival: Survival Analysis, Including Penalised Likelihood. R package version 2.36-10, URL http://CRAN.R-project.org/package=survival.
}
\author{ Juan Xiong, Wenqing He and Grace Y. Yi}



