\name{simexaft-package}
\alias{simexaft-package}
\docType{package}
\title{
SIMEX algorithm for the accelerated failure time model with mismeasured covariates
}
\description{
Implementation of Simulation-Extrapolation (SIMEX) algorithm for the accelerated failure time (AFT) model with mismeasured covariates.
}

\details{
\tabular{ll}{
Package: \tab simexaft\cr
Type: \tab Package\cr
Version: \tab 1.0.7\cr
Date: \tab 2014-01-19\cr
License: \tab GPL\cr
Imports: \tab mvtnorm, survival\cr
LazyLoad: \tab yes\cr
}
}
\author{
Juan Xiong <xiongjuan2000@gmail.com>, Wenqing He <whe@stats.uwo.ca>, Grace Y. Yi<yyi@uwaterloo.ca>

Maintainer: Juan Xiong <xiongjuan2000@gmail.com>
}

\references{
Genz, A., Bretz, F., Miwa, T., Mi, X., Leisch, F., Scheipl, F. and Hothorn, T. (2011). mvtnorm: Multivariate Normal and t Distributions. R package version 0.9-9991, URL http://CRAN.
R-project.org/package=mvtnorm.

He, W., Yi, G. Y. and Xiong, J. (2007). Accelerated Failure Time Models with Covariates Subject to Measurement Error. \emph{Statistics in Medicine}, 26, 4817-4832.

Therneau, T. and Lumley, T. (2011). survival: Survival Analysis, Including Penalised Likelihood. R package version 2.36-10, URL http://CRAN.R-project.org/package=survival.
}
\keyword{ package }

