\name{print.simexaft}
\alias{print.simexaft}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
 Print Method for the SIMEXAFT Class
}
\description{
Printing the most important values in a clear way. 
}
\usage{
\method{print}{simexaft}(x, digits = max(3, getOption("digits") - 3), \dots)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ object of class SIEMXAFT. }
  \item{digits}{ number of digits to be printed. }
  \item{\dots}{ arguments passed to other functions. }
}
\references{
He, W., Yi, G. Y. and Xiong, J. (2007). Accelerated Failure Time Models with Covariates Subject to Measurement Error. \emph{Statistics in Medicine}, 26, 4817-4832.
}
\author{ Juan Xiong, Wenqing He and Grace Y. Yi}

