\name{simexaft-package}
\alias{simexaft-package}
\docType{package}
\title{
simexaft
}
\description{
Implementation of Simulation-Extrapolation (SIMEX) algorithm for the accelerated failure time (AFT) model with mismeasured covariates.
}
\details{
\tabular{ll}{
Package: \tab simexaft\cr
Type: \tab Package\cr
Version: \tab 1.0.2\cr
Date: \tab 2010-07-08\cr
License: \tab GPL \cr
LazyLoad: \tab yes\cr
}

}

\author{
Juan Xiong <jxiong2@uwo.ca>, Wenqing He <whe@stats.uwo.ca>, Grace Y. Yi<yyi@uwaterloo.ca>

Maintainer: Juan Xiong <jxiong2@uwo.ca>
}
\references{
He WQ, Yi GY, Xiong J.(2007) Accelerated Failure Time Models with Covariates Subject to Measurement Error. \emph{Statistics in Medicine}, 26, 4817-4832.
}

\keyword{ package }
